requirements.txt
aiohttp==3.6.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
discord.py==1.5.0
dj-database-url==0.5.0
Django==3.1.4
django-heroku==0.3.1
ffmpeg-python==0.2.0
future==0.18.2
gTTS==2.2.1
gunicorn==20.0.4
idna==2.10
multidict==4.7.6
patool==1.12
playsound==1.2.2
psycopg2==2.8.6
pycparser==2.20
pygame==2.0.1
PyNaCl==1.4.0
pytz==2020.5
rarfile==4.0
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
yarl==1.5.1
